Add Water Shader
This adds the water shader. It is based on the sum of sines technique, mentioned in the GPU Gems book. I used a custom plane (made in Blender) to get more vertices for better waves. There is still visible tiling in the water that needs to be worked on at a later stage.