Fix BoxCollider Detector
The box collider used to detect the interactable objects wasn't following the player's rotation. This causes it to trigger even when the player is looking away from the object. As a short term fix, I have changed it to use a fixed collider on the player instead.

Additionally, when the player walks away, the IKTarget would stay on the object, placing the player's arm at a weird angle. To fix this, I added a check to stop the interaction if the player is walking away.